refactor: :truck: organize hook folder
- change type and hook import directory from alias `$lib/...` to reletive path
